Creating cable create component and changing cables list to new component
